refactor: 폭죽 UseCase 60줄 제한 준수 및 포트 의존성 분리
