+++ src/main/kotlin/com/team2/server/common/image/application/port/ImageUrlPort.kt
@@ -1,0 +1,10 @@
+package com.team2.server.common.image.application.port
+
+import com.team2.server.common.image.entity.ImageTargetType
+
+interface ImageUrlPort {
+    fun findFirstImageUrlByTargetIds(
+        targetType: ImageTargetType,
+        targetIds: Collection<Long>,
+    ): Map<Long, String>
+}
