+++ src/main/kotlin/com/team2/server/fireworks/application/dto/FireworksPayload.kt
@@ -1,0 +1,7 @@
+package com.team2.server.fireworks.application.dto
+
+data class FireworksPayload(
+    val partyId: Long,
+    val participantId: Long,
+    val nickname: String,
+)
