+++ src/test/kotlin/com/team2/server/fireworks/application/usecase/TriggerFireworksUseCaseTest.kt
@@ -3,11 +3,8 @@ package com.team2.server.fireworks.application.usecase
 import com.team2.server.chat.application.port.PartySseEventPublisher
 import com.team2.server.common.exception.BusinessException
 import com.team2.server.common.exception.ErrorCode
-import com.team2.server.common.image.entity.ImageTargetType
-import com.team2.server.common.image.persistence.ImageUrlReader
 import com.team2.server.party.application.usecase.ResolveLiveOpenRealtimePartyUseCase
 import com.team2.server.party.application.usecase.ResolveRealtimeParticipantProfileUseCase
-import com.team2.server.party.domain.entity.Character
 import com.team2.server.party.domain.entity.Participant
 import com.team2.server.party.domain.entity.RealtimeParticipantProfile
 import com.team2.server.party.domain.entity.RealtimeParty
@@ -23,9 +20,6 @@ import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import java.time.Clock
-import java.time.Instant
-import java.time.ZoneId
 import kotlin.test.assertEquals
 
 @ExtendWith(MockitoExtension::class)
@@ -34,11 +28,7 @@ class TriggerFireworksUseCaseTest {
 
     @Mock lateinit var resolveRealtimeParticipantProfileUseCase: ResolveRealtimeParticipantProfileUseCase
 
-    @Mock lateinit var imageUrlReader: ImageUrlReader
-
     @Mock lateinit var partySseEventPublisher: PartySseEventPublisher
-
-    private val fixedClock: Clock = Clock.fixed(Instant.parse("2026-05-29T10:00:00Z"), ZoneId.of("Asia/Seoul"))
 
     private lateinit var useCase: TriggerFireworksUseCase
 
@@ -48,9 +38,7 @@ class TriggerFireworksUseCaseTest {
             TriggerFireworksUseCase(
                 resolveLiveOpenRealtimePartyUseCase = resolveLiveOpenRealtimePartyUseCase,
                 resolveRealtimeParticipantProfileUseCase = resolveRealtimeParticipantProfileUseCase,
-                imageUrlReader = imageUrlReader,
                 partySseEventPublisher = partySseEventPublisher,
-                clock = fixedClock,
             )
     }
 
@@ -59,17 +47,13 @@ class TriggerFireworksUseCaseTest {
     private fun makeProfile(
         participantId: Long = 1L,
         nickname: String = "토끼왕",
-        isCelebrant: Boolean = false,
-        character: Character? = null,
     ): RealtimeParticipantProfile {
         val participant: Participant = mock()
         whenever(participant.id).thenReturn(participantId)
-        whenever(participant.isCelebrant).thenReturn(isCelebrant)
 
         val profile: RealtimeParticipantProfile = mock()
         whenever(profile.participant).thenReturn(participant)
         whenever(profile.nickname).thenReturn(nickname)
-        whenever(profile.character).thenReturn(character)
         return profile
     }
 
@@ -94,35 +78,7 @@ class TriggerFireworksUseCaseTest {
 
     @Test
     fun `참가자가 폭죽을 트리거하면 SSE fireworks 이벤트가 broadcast된다`() {
-        val character: Character = mock()
-        whenever(character.id).thenReturn(2L)
-        val profile = makeProfile(participantId = 5L, nickname = "토끼왕", isCelebrant = false, character = character)
-
-        whenever(resolveLiveOpenRealtimePartyUseCase.invoke(10L)).thenReturn(makeParty())
-        whenever(resolveRealtimeParticipantProfileUseCase.invoke(10L, null, "tok")).thenReturn(profile)
-        whenever(imageUrlReader.findFirstImageUrlByTargetIds(eq(ImageTargetType.CHARACTER), eq(listOf(2L))))
-            .thenReturn(mapOf(2L to "https://cdn.example.com/char2.png"))
-
-        useCase.invoke(10L, null, "tok")
-
-        verify(partySseEventPublisher).broadcastAfterCommit(eq(10L), any(), anyOrNull())
-    }
-
-    @Test
-    fun `캐릭터 없는 참가자도 broadcast된다`() {
-        val profile = makeProfile(participantId = 5L, nickname = "익명", character = null)
-
-        whenever(resolveLiveOpenRealtimePartyUseCase.invoke(10L)).thenReturn(makeParty())
-        whenever(resolveRealtimeParticipantProfileUseCase.invoke(10L, null, "tok")).thenReturn(profile)
-
-        useCase.invoke(10L, null, "tok")
-
-        verify(partySseEventPublisher).broadcastAfterCommit(eq(10L), any(), anyOrNull())
-    }
-
-    @Test
-    fun `주최자(CELEBRANT)가 트리거해도 broadcast된다`() {
-        val profile = makeProfile(participantId = 1L, nickname = "주최자", isCelebrant = true)
+        val profile = makeProfile(participantId = 5L, nickname = "토끼왕")
 
         whenever(resolveLiveOpenRealtimePartyUseCase.invoke(10L)).thenReturn(makeParty())
         whenever(resolveRealtimeParticipantProfileUseCase.invoke(10L, null, "tok")).thenReturn(profile)
